Sync with upstream @ 3dd8ca09 (#42)
* modified sub headings for clarity and to avoid confusion

* Update article.md

* Update task.md

fix typo

* Update article.md

Change to wording about maximum recursion depth.

* Update article.md

* Update article.md

Correct spelling of "standardized".

* Fix typo

* minor

* closes #1271

* minor

* clarify type conversion headings

* minor

* improve description

Change the order between ´body´ and ´step´. Also delete the phrase '... but before the condition check', which seems a bit confusing.

* Update article.md

Maybe this is what you want. "Complexity" is singular, so if that's the antecedent, one would use "was".

* fixes #1261

* Fix typo

fix typo in comment to match code example

* Typo/grammar fix

* fixed missing pronoun 'it'

fixed missing pronoun 'it'.

* fixed missing pronoun 'it'

fixed missing pronoun 'it'

* Update article.md

fixed grammaticalmistakes, i.e. changed 'And' to 'and', added missing pronoun 'it' after 'to'.

* Update article.md

* Fixed spelling error

So here JavaScript assumes that we have a code block, but why there's an error. --> Makes no sense so replaced "but" with "that's"

* fixes

* minor

* Update article.md

Better grammatically.

* typo

* minor

* fixes

* fix minor English language issues

* fix typo

* Some minor corrections

* Is it

* Update article.md

Wording of initial explanation of spread operator.

* Typo

* there must be exactly the origin there.

* Fixed a typo

* Update article.md

1. Since is general statement, thought might be better to use general function name.
2. "Do the following steps" is a bit awkward, in that usually one doesn't "do" steps (but rather "takes", or "follows", or "executes" steps, perhaps). But it does convey the meaning, and perhaps it's a relatively new usage in business or tech that I'm not familiar with. For me, it comes off a bit awkward. Leaving out "steps" works grammatically and preserves the sense. Or could find another verb to replace "do" -- but nothing perfect came to mind. Not that bad the way it is, but since I was already changing the sentence...

* Update article.md

Says "is the same object as:", but what's below is an assignment (that includes an object). I didn't have any problem understanding it, didn't even notice it at first, but I thought it would be more accurate this way without sacrificing clarity.

* Update article.md

Qualify that while cachingDecorator() will now work with any number of arguments, the hash() function given will not. Or, perhaps better, could just change the hash function to something like:

function hash(args) {
  return [...args].join();
}

* Update article.md

* minor

* fixes

* grammar fix

removed 'is'

* minor

* minor

* minor

* minor

* minor

* minor

* minor

* minor

* closes #1310

* closes #1309

* closes #1309

* closes #1309

* minor

* closes #1308

* closes #1294

* Fix typo in Proxy and Reflect article

* Update task.md

Wording was ungrammatical and hard to understand. I made my best guess as to what you want to say. However, it is not clear from the description or the examples whether the debounce wrapper should inhibit calls attempted ms milliseconds after the last call that wasn't ignored, or ms milliseconds after the last call that was attempted, even if it was ignored. I would think it would be the former, but the example could go either way. If the f(4) example used 1000 or 1001 ms and still runs, that would clear it up. If it has to be 1100 like it is, that would suggest that it's the second scenario above. In my propsed change I preserved the ambiguity.

* closes #1220

* Update MDN docs link for tagged templates

* Update article.md

As per issue comment thread #1282: 
Clarifying "current code" for zero delay setInterval. 
Plus minor edit of ordering of terms for greater consistency.
Plus minor grammatical fixes.

* an -> a

* WIP

* fixes #1316

* closes #1318

* minor

* minor

* WIP

* WIP

* regexp

* WIP

* reg->regexp

* WIP

* Typo fix

* minor

* form

* Typo fix

* fixes

* fixes

* fix

* Fix minor grammatical errors

* Minor grammatical fixes

* minor